Refactor: rename RabbitConsumers folders to the Consumers
