+++ Identity/Hotel.Identity.Infrastructure/ServiceCollectionExtensions.cs
@@ -1,8 +1,8 @@
 using Hotel.Identity.Domain.Abstractions;
 using Hotel.Identity.Infrastructure.Abstractions;
+using Hotel.Identity.Infrastructure.Consumers;
 using Hotel.Identity.Infrastructure.Database;
 using Hotel.Identity.Infrastructure.Models;
-using Hotel.Identity.Infrastructure.RabbitConsumers;
 using Hotel.Identity.Infrastructure.Services;
 using MassTransit;
 using Microsoft.EntityFrameworkCore;
